corrected data type molecular weight in dto of product and reagent

+++ src/main/java/ru/itpark/larec/dto/ProductDto.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 public class ProductDto {
     private int experimentId;
     private String name;
-    private int molecularWeight;
+    private BigDecimal molecularWeight;
     private String formula;
     private String linkForDescription;
     private BigDecimal amountGram;
